requirements.txt
esptool==4.7.0
pyserial==3.5
tkinter-tooltip==2.0.0
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
requests>=2.25.0
websocket-client==1.6.4
typing-extensions>=3.7.4
